+++ demos/pagefind.ts
@@ -5,15 +5,16 @@ const binfile = await dbin({
     "https://github.com/CloudCannon/pagefind/releases/download/{version}/pagefind-{version}-{target}.tar.gz",
   checksumPattern:
     "https://github.com/CloudCannon/pagefind/releases/download/{version}/pagefind-{version}-{target}.tar.gz.sha256",
-  version: "v0.8.1",
+  version: "v1.3.0",
   targets: [
     { name: "x86_64-unknown-linux-musl", os: "linux", arch: "x86_64" },
-    { name: "x86_64-apple-darwin", os: "darwin" },
+    { name: "aarch64-apple-darwin", os: "darwin", arch: "aarch64" },
     { name: "x86_64-pc-windows-msvc", os: "windows", arch: "x86_64" },
   ],
   dest: "./_bin/pagefind",
 });
 
-const process = Deno.run({ cmd: [binfile, "-h"] });
-await process.status();
-process.close();
+new Deno.Command(binfile, {
+  stdout: "inherit",
+  args: ["-h"],
+}).outputSync();